pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi>=0.128.0",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "uvicorn>=0.40.0",
    "PyJWT>=2.8.0",
    "httpx>=0.28.0",
    "fpdf2>=2.8.0"
]
